Conflicts:
	NDDigital.DiarioAcademia.Aplicacao/Services/AuthorizationService.cs

	NDDigital.DiarioAcademia.Aplicacao/Services/Security/IPermissionService.cs
	NDDigital.DiarioAcademia.Infraestrutura.Orm/Security/MyUserStore.cs

+++ NDDigital.DiarioAcademia.WebApi/Controllers/Authentication/PermissionController.cs
@@ -1,5 +1,9 @@
-﻿using NDDigital.DiarioAcademia.Aplicacao.Services;
+﻿using Infrastructure.DAO.ORM.Common;
+using NDDigital.DiarioAcademia.Aplicacao.Services;
 using NDDigital.DiarioAcademia.Dominio.Entities.Security;
+using NDDigital.DiarioAcademia.Infraestrutura.DAO.Common.Factorys;
+using NDDigital.DiarioAcademia.Infraestrutura.DAO.Common.Uow;
+using NDDigital.DiarioAcademia.Infraestrutura.IoC;
 using NDDigital.DiarioAcademia.Infraestrutura.Orm.Common;
 using NDDigital.DiarioAcademia.Infraestrutura.Orm.Security;
 using System.Threading.Tasks;
@@ -11,16 +15,15 @@ namespace NDDigital.DiarioAcademia.WebApi.Controllers.Authentication
     {
         IPermissionService _permissionService;
 
-        public PermissionController()
+        public PermissionController() //TODO: IOC
         {
+            var factory = new EntityFrameworkFactory();
 
-            var factory = new DatabaseFactory();
+            var unitOfWork = new EntityFrameworkUnitOfWork(factory);
 
-            var uow = new UnitOfWork(factory);
+            var permissionRepo = new PermissionRepository(factory); //Container.Get<IPermissionRepository>();
 
-            var permissionRepo = new PermissionRepository(factory);
-
-            _permissionService = new PermissionService(permissionRepo, uow);
+            _permissionService = new PermissionService(permissionRepo, unitOfWork);
 
         }
 
@@ -36,20 +39,29 @@ namespace NDDigital.DiarioAcademia.WebApi.Controllers.Authentication
             return Ok(_permissionService.GetById(id));
         }
 
-        // POST: api/Permission
-        public IHttpActionResult Post([FromBody]Permission value)
+        public IHttpActionResult Post([FromBody]Permission[] values)
         {
-            _permissionService.Add(value);
-            return Ok(value);
+
+            foreach (var item in values)
+            
+               _permissionService.Add(item);
+            
+            return Ok(values);
         }
 
-        // DELETE: api/Permission/5
-        public void Delete(string id)
+        // DELETE: api/Permission/
+        public IHttpActionResult Delete([FromBody]string[] ids)
         {
-            var permission = _permissionService.GetByPermissionId(id);
+            if(ids == null)
+                return BadRequest();
+            foreach (var id in ids)
+            {
+                var permission = _permissionService.GetByPermissionId(id);
 
-            _permissionService.Delete(permission.Id);
+                _permissionService.Delete(permission.Id);
 
+            }
+            return Ok();
         }
     }
 }
